Rearrange CLI channel command
Possibly a little deep of a command, but it isn't common
