+++ crates/fluvio-cli/src/bin/main.rs
@@ -3,7 +3,7 @@ use color_eyre::eyre::Result;
 use fluvio_cli::{Root, HelpOpt};
 use fluvio_future::task::run_block_on;
 #[cfg(not(target_os = "windows"))]
-use fluvio_cli::cli_config::{CliChannelName, FluvioChannelConfig, is_fluvio_bin_in_std_dir};
+use fluvio_cli::cli_config::channel::{CliChannelName, FluvioChannelConfig, is_fluvio_bin_in_std_dir};
 #[cfg(not(target_os = "windows"))]
 use std::process::Stdio;
 #[cfg(not(target_os = "windows"))]
