+++ crates/fluvio-cli/src/install/plugins.rs
@@ -1,6 +1,6 @@
 use structopt::StructOpt;
 use fluvio_index::{PackageId, HttpAgent, MaybeVersion};
-use crate::cli_config::{CliChannelName, FluvioChannelConfig};
+use crate::cli_config::channel::{CliChannelName, FluvioChannelConfig};
 
 use crate::{Result, CliError};
 use crate::install::{
@@ -124,10 +124,6 @@ impl InstallOpt {
         let fluvio_dir = fluvio_extensions_dir()?;
         let package_filename = if target.to_string().contains("windows") {
             format!("{}.exe", id.name().as_str())
-        } else if channel.current_channel() == CliChannelName::Latest {
-            let mut latest_bin = id.name().to_string();
-            latest_bin.push_str("-latest");
-            latest_bin
         } else {
             id.name().to_string()
         };
